+++ sync-for-payables/src/codatsyncpayables/bill_payments.py
@@ -29,12 +29,12 @@ class BillPayments:
         
         url = utils.generate_url(operations.CreateBillPaymentRequest, base_url, '/companies/{companyId}/connections/{connectionId}/push/billPayments', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "bill_payment", 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, "bill_payment", True, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         query_params = utils.get_query_params(operations.CreateBillPaymentRequest, request)
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -48,7 +48,7 @@ class BillPayments:
 
         def do_request():
             return client.request('POST', url, params=query_params, data=data, files=form, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',
@@ -83,8 +83,8 @@ class BillPayments:
         ### Process
         1. Pass the `{billPaymentId}` to the *Delete bill payment* endpoint and store the `pushOperationKey` returned.
         2. Check the status of the delete operation by checking the status of push operation either via
-            1. [Push operation webhook](https://docs.codat.io/introduction/webhooks/core-rules-types#push-operation-status-has-changed) (advised),
-            2. [Push operation status endpoint](https://docs.codat.io/sync-for-payables-api#/operations/get-push-operation).
+           1. [Push operation webhook](https://docs.codat.io/introduction/webhooks/core-rules-types#push-operation-status-has-changed) (advised),
+           2. [Push operation status endpoint](https://docs.codat.io/sync-for-payables-api#/operations/get-push-operation).
 
            A `Success` status indicates that the bill payment object was deleted from the accounting platform.
         3. (Optional) Check that the bill payment was deleted from the accounting platform.
@@ -95,16 +95,22 @@ class BillPayments:
         ## Integration specifics
         Integrations that support soft delete do not permanently delete the object in the accounting platform.
 
-        | Integration | Soft Delete | Details                                                                                             |  
-        |-------------|-------------|-----------------------------------------------------------------------------------------------------|
-        | Oracle NetSuite   | No          | See [here](/integrations/accounting/netsuite/accounting-netsuite-how-deleting-bill-payments-works) to learn more. |
+        | Integration | Soft Delete | Details                                                                                              |  
+        |-------------|-------------|------------------------------------------------------------------------------------------------------|                                                        
+        | Oracle NetSuite   | No          | See [here](/integrations/accounting/netsuite/how-deleting-bill-payments-works) to learn more.  |
+        | QuickBooks Online | No          | -                                                                                              |
+        | Xero | Yes          | -                                                                                                          |
+
+        > **Supported integrations**
+        >
+        > This functionality is currently supported for our QuickBooks Online, Xero and Oracle NetSuite integrations.
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())
         
         url = utils.generate_url(operations.DeleteBillPaymentRequest, base_url, '/companies/{companyId}/data/billPayments/{billPaymentId}', request)
         headers = {}
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -118,7 +124,7 @@ class BillPayments:
 
         def do_request():
             return client.request('DELETE', url, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',
@@ -159,7 +165,7 @@ class BillPayments:
         url = utils.generate_url(operations.GetBillPaymentsRequest, base_url, '/companies/{companyId}/data/billPayments/{billPaymentId}', request)
         headers = {}
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -173,7 +179,7 @@ class BillPayments:
 
         def do_request():
             return client.request('GET', url, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',
@@ -216,7 +222,7 @@ class BillPayments:
         url = utils.generate_url(operations.GetCreateBillPaymentModelRequest, base_url, '/companies/{companyId}/connections/{connectionId}/options/billPayments', request)
         headers = {}
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -230,7 +236,7 @@ class BillPayments:
 
         def do_request():
             return client.request('GET', url, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',
@@ -270,7 +276,7 @@ class BillPayments:
         headers = {}
         query_params = utils.get_query_params(operations.ListBillPaymentsRequest, request)
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -284,7 +290,7 @@ class BillPayments:
 
         def do_request():
             return client.request('GET', url, params=query_params, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',