+++ sync-for-payables/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [python v1.3.1] sync-for-payables
 ### Releases
 - [PyPI v1.3.1] https://pypi.org/project/codat-sync-for-payables/1.3.1 - sync-for-payables
+
+## 2023-09-29 00:17:30
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.4.0] sync-for-payables
+### Releases
+- [PyPI v1.4.0] https://pypi.org/project/codat-sync-for-payables/1.4.0 - sync-for-payables